FEAT: Deployed production package to AWS Elastic Beanstalk
-Minor bug fixes identified by spotbugs

+++ src/main/java/edu/bu/cs673/stockportfolio/api/portfolio/PortfolioController.java
@@ -69,10 +69,14 @@ public class PortfolioController {
     public String deletePortfolio(Authentication authentication, Model model) {
         User currentUser = getCurrentUser(authentication);
         Portfolio currentPortfolio = currentUser.getPortfolio();
-        Long id = currentPortfolio.getId();
+
+        if (currentPortfolio == null) {
+            return "home";
+        }
 
         boolean result = validationService.validatePortfolioOwner(currentPortfolio, currentUser, model, "delete");
 
+        Long id = currentPortfolio.getId();
         if (result) {
             portfolioService.deletePortfolioBy(id);
         }
